Update MMD Tools to latest Alpha for 4.2.
