--- extern_tools/mmd_tools_local/CHANGELOG.md
@@ -1,162 +1,0 @@
-変更履歴
-========
-
-See the following URL for a recent changelog.
-* https://github.com/UuuNyaa/blender_mmd_tools/releases
-
-最近の変更履歴は以下のURLを参照してください。
-* https://github.com/UuuNyaa/blender_mmd_tools/releases
-
-
-
-
-v0.4.5
-------
-### 機能追加 ###
-* スフィアマップ(sph, spa)に対応
-    * ブレンド率をモデルインポート画面にて設定できます。
-* モデルインポート画面にミップマップのオンオフチェックボックスを追加
-* モデルのインポート速度を改善
-
-### 変更 ###
-* 各マテリアルの *Receive Transparent* をONに設定
-
-### バグ修正 ###
-* アルファチャンネルを持つテクスチャを使用したモデルが正常にインポートされない不具合を修正
-* sph テクスチャが適切なテクスチャ座標で読み込まれない不具合を修正
-
-
-v0.4.4
-------
-### バグ修正 ###
-* プラグインアンロードが正常に行われない不具合を修正
-
-
-v0.4.3
-------
-### 機能追加 ###
-* MMD Rig Toolsパネルを追加
-    * 剛体、ジョイント等の表示/非表示を切り替えることができます。
-
-### 変更 ###
-* "distance of ignore collisions"の精度を向上
-    * 非衝突グループを無視する距離を剛体のサイズに応じて動的に最適化するよう変更
-    * 初音ミクVer2.pmdで物理シミュレーションが破綻する問題が直ります。
-
-v0.4.2
-------
-### 変更 ###
-* インポート速度を向上
-* "distance of ignore collisions"のデフォルト値を3.0に変更
-
-v0.4.1
-------
-### 機能追加 ###
-* インポート画面にログレベルを追加
-
-### バグ修正 ###
-* インポートに失敗することがある不具合を修正
-    * 正常に読み込めているモデルがたまに読み込めない現象の修正
-
-v0.4.0
-------
-### 機能追加 ###
-* pmdファイルのインポートに対応
-
-### バグ修正 ###
-* 剛体を持たないモデルのみをインポートしているシーンにvmdを読み込むと
-  フレームレンジ設定でエラーが発生する不具合を修正
-
-v0.3.3
-------
-### バグ修正 ###
-* 回転付与ボーンが連鎖する場合、回転付与結果がMMDと大幅に異なる問題に対応
-    * 回転付与ボーンの付与親が回転付与ボーンだった場合の処理を変更
-    * 「Tda式初音ミク・アペンド」の"目"が正しく動作しない問題が解決します
-        * 作者: Tdaさん[HP](http://u16hp.blog134.fc2.com/)
-
-v0.3.2
-------
-### 機能追加 ###
-* CyclesとBlender renderの相互切り替え機能を追加
-    * ただし、マテリアルの設定は共有されません
-
-### 変更 ###
-* "To cycles"をGLSL View機能と対等な機能として再実装
-
-### バグ修正 ###
-* 「移動付与」ボーンを持つpmxを読み込めない不具合を修正
-* 「移動付与」ボーンが正常に動作しない不具合をREADME.mdの「既知の問題」に追加
-* IKボーンとターゲットボーンが離れている場合、Blender内での挙動がMMDと異なる問題に対応
-
-v0.3.1
-------
-### 機能追加 ###
-* To cycles時にAlphaを考慮したシェーダを生成するよう変更
-
-### バグ修正 ###
-* mmd_tools_local_local v0.3.0未満で読み込んだモデルへモーションをアサインできるよう修正
-
-v0.3.0
--------
-### 変更 ###
-* Blender addon名を"mmd_tools_local_local"に変更
-* pmxインポート時の"delete tip bones"機能を廃止
-    * 別機能として実装予定
-* pmxインポート時に自動設定されるgravity値の基準を変更
-* PoseBoneのカスタムプロパティ名を変更
-    * isTipBone -> is\_mmd\_tip\_bone
-    * name_j -> mmd\_bone\_name\_j
-    * name_e -> mmd\_bone\_name\_e
-
-### 機能追加 ###
-* "GLSL View"機能を追加
-    * ボタン一つでGLSL表示モードに変更
-    * Resetボタンで初期状態に復帰
-    * "Shadeless"ボタンでshadeless表示
-* "Set frame range"機能を追加
-    * フレームレンジを現在読み込まれているアニメーションの長さに設定
-    * フレームレートを30fpsに変更
-* "Shadeless"機能を追加
-    * 選択中メッシュのマテリアルを全て"Shadeless"に変更
-* ツールパネルを変更
-    * "Import pmx", "Import vmd"ボタンをそれぞれ"Model", "_Motion"ボタンに変更
-    * "auto setup scene"を廃止し、"Set frame range"ボタンを追加
-    * "Reset", "GLSL"ボタンを追加
-    * "Shadeless"ボタンを追加
-* vmdインポート機能に"update scne settings"オプションを追加
-* Alpha値を持つテクスチャに対応
-* ボーンの「回転付与」「移動付与」の実現方法を変更（よりMMDに近い結果になるよう変更）
-* IKのtargetが自身の子ボーンだった場合、そのボーンを自身の親に養子に出す処理を追加（DIVA風ミクver105改.pmx対応）
-* "To cycles"機能をblender 2.67に対応
-* blenderの対応バージョンを2.67以降に変更
-
-### バグ修正 ###
-* ボーン名が重複した場合、正常に「回転付与」「移動付与」が適用されない不具合を修正
-
-v0.2.2
--------
-### 機能追加 ###
-* ボーンの「回転付与」「移動付与」に対応
-* ローカル軸の情報をカスタムプロパティとして読み込む機能を追加
-
-v0.2.1
--------
-### バグ修正 ###
-* vmdからカメラおよびライトモーションをインポートする際に作成するEmptyオブジェクトの名前を変更
-* pmxの読み込み時のスケールを0.2以外の値にすると、物理シミュレーションが破綻する不具合を修正
-
-v0.2.0
-------
-### 機能追加 ###
-* 剛体インポート機能の追加
-
-v0.1.1
--------
-### 機能追加およびバグ修正 ###
-* iRi-Eさんのforkから変更をマージ
-* ・一部読み込めなかったvmdに対応など
-
-v0.1.0
--------
-* テストリリース
