--- extern_tools/mmd_tools_local/CONTRIBUTORS.md
@@ -1,21 +1,0 @@
-Original author
-- sugiany
-
-Former Maintainer
-- powroupi
-
-Maintainer
-- UuuNyaa
-
-Contributors
-- iRi-E
-- lordscales91
-- Pixiy
-- r1cebank
-- nagadomi
-- ptrthomas
-- stefnotch
-- qwertychouskie
-- ChennyBaBy
-- Darkblader24
-- aoowweenn
