--- extern_tools/mmd_tools_local/README.md
@@ -1,24 +1,0 @@
-# MMD Tools
-MMD Tools is a blender addon for importing MMD (MikuMikuDance) model data (.pmd, .pmx), motion data (.vmd) and pose data (.vpd). Exporting model data (.pmx), motion data (.vmd) and pose data (.vpd) are supported as well.
-
-MMD ToolsはMMD(MikuMikuDance)のモデルデータ(.pmd, .pmx)、モーションデータ(.vmd)、ポーズデータ(.vpd)をインポートするためのBlenderアドオンです。モデルデータ(.pmx)、モーションデータ(.vmd)、ポーズデータ(.vpd)のエクスポートにも対応しています。
-
-## Installation & Usage
-- Check [the MMD Tools Wiki](https://mmd-blender.fandom.com/wiki/mmd_tools) for details.
-- 詳細は [MMD ToolsのWiki (日本語)](https://mmd-blender.fandom.com/ja/wiki/mmd_tools) を確認してください。
-
-## Contributing
-MMD Tools needs your contribution!
-
-- Document writing / translation
-- Video creation / translation
-- Bug reports
-- Feature requests
-- Pull requests
-
-If you are interested in any of the above, please contact me via the following channels:
-- [MMD Tools Issues](https://github.com/UuuNyaa/blender_mmd_tools/issues)
-- [MMD & Blender Discord Server](https://discord.gg/zRgUkuaPWw)
-
-## License
-Distributed under the [GPLv3](LICENSE) License.
