+++ extern_tools/mmd_tools_local/core/model.py
@@ -13,7 +13,7 @@ import rna_prop_ui
 from mathutils import Vector
 
 from mmd_tools_local import mmd_tools_local_VERSION, bpyutils
-from mmd_tools_local.bpyutils import FnContext, Props, SceneOp
+from mmd_tools_local.bpyutils import FnContext, Props
 from mmd_tools_local.core import rigid_body
 from mmd_tools_local.core.morph import FnMorph
 from mmd_tools_local.core.rigid_body import MODE_DYNAMIC, MODE_DYNAMIC_BONE, MODE_STATIC
@@ -571,50 +571,49 @@ class Model:
         self.__temporary_grp: Optional[bpy.types.Object] = None
 
     @staticmethod
-    def create(name, name_e="", scale=1, obj_name=None, armature=None, add_root_bone=False):
-        scene = SceneOp(bpy.context)
+    def create(name: str, name_e: str = "", scale: float = 1, obj_name: Optional[str] = None, armature_object: Optional[bpy.types.Object] = None, add_root_bone: bool = False):
         if obj_name is None:
             obj_name = name
 
-        root = bpy.data.objects.new(name=obj_name, object_data=None)
+        context = FnContext.ensure_context()
+
+        root: bpy.types.Object = bpy.data.objects.new(name=obj_name, object_data=None)
         root.mmd_type = "ROOT"
         root.mmd_root.name = name
         root.mmd_root.name_e = name_e
         root["mmd_tools_local_version"] = mmd_tools_local_VERSION
         setattr(root, Props.empty_display_size, scale / 0.2)
-        scene.link_object(root)
+        FnContext.link_object(context, root)
 
-        armObj = armature
-        if armObj:
-            m = armObj.matrix_world
-            armObj.parent_type = "OBJECT"
-            armObj.parent = root
-            # armObj.matrix_world = m
+        if armature_object:
+            m = armature_object.matrix_world
+            armature_object.parent_type = "OBJECT"
+            armature_object.parent = root
+            # armature_object.matrix_world = m
             root.matrix_world = m
-            armObj.matrix_local.identity()
+            armature_object.matrix_local.identity()
         else:
-            arm = bpy.data.armatures.new(name=obj_name)
-            armObj = bpy.data.objects.new(name=obj_name + "_arm", object_data=arm)
-            armObj.parent = root
-            scene.link_object(armObj)
-        armObj.lock_rotation = armObj.lock_location = armObj.lock_scale = [True, True, True]
-        setattr(armObj, Props.show_in_front, True)
-        setattr(armObj, Props.display_type, "WIRE")
+            armature_object = bpy.data.objects.new(name=obj_name + "_arm", object_data=bpy.data.armatures.new(name=obj_name))
+            armature_object.parent = root
+            FnContext.link_object(context, armature_object)
+        armature_object.lock_rotation = armature_object.lock_location = armature_object.lock_scale = [True, True, True]
+        setattr(armature_object, Props.show_in_front, True)
+        setattr(armature_object, Props.display_type, "WIRE")
 
         from mmd_tools_local.core.bone import FnBone
 
-        FnBone.setup_special_bone_collections(armObj)
+        FnBone.setup_special_bone_collections(armature_object)
 
         if add_root_bone:
             bone_name = "全ての親"
-            with bpyutils.edit_object(armObj) as data:
+            with bpyutils.edit_object(armature_object) as data:
                 bone = data.edit_bones.new(name=bone_name)
                 bone.head = [0.0, 0.0, 0.0]
                 bone.tail = [0.0, 0.0, getattr(root, Props.empty_display_size)]
-            armObj.pose.bones[bone_name].mmd_bone.name_j = bone_name
-            armObj.pose.bones[bone_name].mmd_bone.name_e = "Root"
+            armature_object.pose.bones[bone_name].mmd_bone.name_j = bone_name
+            armature_object.pose.bones[bone_name].mmd_bone.name_e = "Root"
 
-        bpyutils.select_object(root)
+        FnContext.set_active_and_select_single_object(context, root)
         return Model(root)
 
     @staticmethod
@@ -672,7 +671,7 @@ class Model:
             self.__rigid_grp = FnModel.find_rigid_group_object(self.__root)
             if self.__rigid_grp is None:
                 rigids = bpy.data.objects.new(name="rigidbodies", object_data=None)
-                SceneOp(bpy.context).link_object(rigids)
+                FnContext.link_object(FnContext.ensure_context(), rigids)
                 rigids.mmd_type = "RIGID_GRP_OBJ"
                 rigids.parent = self.__root
                 rigids.hide_set(True)
@@ -689,7 +688,7 @@ class Model:
             self.__joint_grp = FnModel.find_joint_group_object(self.__root)
             if self.__joint_grp is None:
                 joints = bpy.data.objects.new(name="joints", object_data=None)
-                SceneOp(bpy.context).link_object(joints)
+                FnContext.link_object(FnContext.ensure_context(), joints)
                 joints.mmd_type = "JOINT_GRP_OBJ"
                 joints.parent = self.__root
                 joints.hide_set(True)
@@ -706,7 +705,7 @@ class Model:
             self.__temporary_grp = FnModel.find_temporary_group_object(self.__root)
             if self.__temporary_grp is None:
                 temporarys = bpy.data.objects.new(name="temporary", object_data=None)
-                SceneOp(bpy.context).link_object(temporarys)
+                FnContext.link_object(FnContext.ensure_context(), temporarys)
                 temporarys.mmd_type = "TEMPORARY_GRP_OBJ"
                 temporarys.parent = self.__root
                 temporarys.hide_set(True)
@@ -1017,7 +1016,7 @@ class Model:
 
                 if "mmd_tools_local_rigid_track" not in target_bone.constraints:
                     empty = bpy.data.objects.new(name="mmd_bonetrack", object_data=None)
-                    SceneOp(bpy.context).link_object(empty)
+                    FnContext.link_object(FnContext.ensure_context(), empty)
                     empty.matrix_world = target_bone.matrix
                     setattr(empty, Props.empty_display_type, "ARROWS")
                     setattr(empty, Props.empty_display_size, 0.1 * getattr(self.__root, Props.empty_display_size))
