+++ extern_tools/mmd_tools_local/operators/camera.py
@@ -5,6 +5,7 @@
 from bpy.props import BoolProperty, EnumProperty, FloatProperty
 from bpy.types import Operator
 
+from mmd_tools_local.bpyutils import FnContext
 from mmd_tools_local.core.camera import MMDCamera
 
 
@@ -54,15 +55,13 @@ class ConvertToMMDCamera(Operator):
 
     def execute(self, context):
         if self.bake_animation:
-            from mmd_tools_local.bpyutils import SceneOp
-
             obj = context.active_object
             targets = [x for x in context.selected_objects if x != obj]
             target = targets[0] if len(targets) == 1 else None
             if self.camera_source == "SCENE":
                 obj = None
             camera = MMDCamera.newMMDCameraAnimation(obj, target, self.scale, self.min_distance).camera()
-            SceneOp(context).active_object = camera
+            FnContext.set_active_object(context, camera)
         else:
             MMDCamera.convertToMMDCamera(context.active_object, self.scale)
         return {"FINISHED"}
