--- extern_tools/mmd_tools_local/typings/mmd_tools/properties/morph.pyi
@@ -1,18 +1,0 @@
-# -*- coding: utf-8 -*-
-# Copyright 2024 MMD Tools authors
-# This file is part of MMD Tools.
-
-class MaterialMorph:
-    pass
-
-class UVMorph:
-    pass
-
-class BoneMorph:
-    pass
-
-class GroupMorph:
-    pass
-
-class VertexMorph:
-    pass
