--- extern_tools/mmd_tools_local/typings/mmd_tools/properties/pose_bone.pyi
@@ -1,26 +1,0 @@
-# -*- coding: utf-8 -*-
-# Copyright 2024 MMD Tools authors
-# This file is part of MMD Tools.
-
-from mathutils import Vector
-
-class MMDBone:
-    name_j: str
-    name_e: str
-    bone_id: int
-    transform_order: int
-    is_controllable: bool
-    transform_after_dynamics: bool
-    enabled_fixed_axis: bool
-    fixed_axis: Vector
-    enabled_local_axes: bool
-    local_axis_x: Vector
-    local_axis_z: Vector
-    is_tip: bool
-    ik_rotation_constraint: float
-    has_additional_rotation: bool
-    has_additional_location: bool
-    additional_transform_bone: str
-    additional_transform_bone_id: int
-    additional_transform_influence: float
-    is_additional_transform_dirty: bool
