+++ ui/settings_updates.py
@@ -26,8 +26,6 @@ class UpdaterPanel(ToolPanel, bpy.types.Panel):
         col.separator()
 
         row = col.row(align=True)
-        row.prop(context.scene, 'show_mmd_tabs')
-        row = col.row(align=True)
         row.prop(context.scene, 'embed_textures')
         row = col.row(align=True)
         row.prop(context.scene, 'remove_rigidbodies_joints_global')
